Simplify battle players
Remove BattleManager and merge it into Client, as it seems two layers of
abstraction are unecessary complex and I do not see a good reason why
Client class could not serve the same purpose as BattleManager, i.e. to
manage and switch various implementations of event providers based on
the app demand.