Update round indicator in arena panel
Replace plain text indicator with more graphical "TV-like" game status
indicator, containing the game type (LIVE/REPLAY), as well as
information about the current round and turn.

Move the indicator to the top left corner of the whole ArenaPanel (not
just battleground) so that it does not interfere with the battleground
view if it does not need to.